[SYSTEMDS-70] Implement lineage tracing for spark instructions
